move is_sm90_supported/is_sm100_supported to python/sglang/srt/utils.py (#9679)
